+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # PyMagento Changelog
 
+## Unreleased
+
+* Rename the parameter `log_progress` to `verbose`. Keep `log_progress` as an alias for backward compatibility.
+
 ## 1.3.4 (2022/05/04)
 
 * Fix a bug in `BatchSaver` where the `batch_size` constructor argument was ignored